Document HAVE_LRINTF in the draft Makefile.
I've changed the grouping of the options to place the performance-
related ones after the more serious thread-safety options. This
is still somewhat confusing as with C99 on wants -DHAVE_LRINTF
but NOT -Drestrict=

I've also added leading whitespace to the comment lines to better
distinguish them from the commented-out code lines.