Add extra entries to the shell code tables
Although the current encoder uses does not use them, the decoder
might have tried to use offsets beyond the end of these tables.
Inserting values here ensures that the stream can be decoded in a
well-defined way.
The particular values of these tables were chosen to match a
Gaussian distribution with sigma=3.93.