Define a fallback version string.
The new opus_get_version_string() call just returns a static string
defined by OPUS_VERSION, which is passed in from the build system
through config.h (or a custom compile line). Provide a fallback to
"unknown" if the build system fails to actually provide that definition.

This restores compilation with Makefile.draft.

In general, this means there will be builds out there with non-specific
version strings, since ports won't be forced to update the string. While
that's unfortunate, I think it's more valuable that the library be simple
to build.