Advertise the static library for uninstalled builds.
It's less surprising for people using the -uninstalled
pkg-config file if we advertise the static library rather
than the shared build. That way the executables work without
having to hack the same location into LD_LIBRARY_PATH, and
the resulting binaries are easier to distribute.

For this to actually work with opus-tools, the -fPIE had
to be turned off, as it was in 1d128b4d5c6f. Support
for static PIE builds would be restored if we compiled
libopus.a with -fPIC.

Also remove the private header locations. These are no
longer required to declare any of the public api.