Fix line endings on MSVC project files.
These files has unix line endings, while the other
project files have dos line endings.

I hoped this would fix the link error about corrupt
or invalid files with these projects, but it doesn't.
Running 'Update project files...' in the VS2012 IDE
is a work around, but it just adds a <PlatformToolset>
element to each <PropertyGroup> and a line ending at
the end of the file.