Update the oggopus script to use xml2rfc v2.
The command line has changed, for the better, but we need to
update the way we call it. Also adds an install hint if it's
not found on the path.

NB I tried to detect and fallback to the old syntax for the
version 1 (TCL) implementation but this is hard because it
doesn't support any switches (like --version) and it no
longer parses our remote entities.