Add float suffix to tansig_table literals.
This fixes a truncation warning in MSVC:

warning C4305: 'initializing' : truncation from 'double' to
'const opus_val16'.

opus_val16 is a float in the floating point build and an
int16 in the fixed-point build. The table isn't used in
the fixed-point build, but the compiler would presumedly
warn in that case as well.