Clarify how grouped streams are not "Ogg Opus files".
We mean to make .opus only for degenerate--not concurrently
multiplexed or grouped--but optionally sequentially multiplexed
or chained physical Ogg streams.

However, we don't mean to say you can't use this draft with
such streams.

Also, replace the reference to RFC3534 with RFC5334, which
obsoletes the previous reference by defining video/ogg and
audio/ogg.