Improve Content Type figure placement.
With xml2rfc 2.3.9, I was getting the 'audio/ogg; codecs=opus'
figure hoisted out of the the containing paragraph, which was
very confusing.

I've tried to improve this by moving the surrounding paragraph
into the figure using the <preamble> and <postamble> tags, as
in the example in draft-mrose-writting-rfcs.

Unfortunately this still isn't perfect. Since the surrounding
paragraph is part of the figure we can't set align="center" on
the figure itself, and the processor seems to ignore it on the
artwork element. I've compensated by adding some leading
whitespace.