Clarify that MB->WB is better than MB->NB
I hope.  Now that it's been clarified for me what this originally meant
to say, it's tricky to know if this actually contains the missing clue
for other readers.  Hopefully it should be obvious to anyone actually
implementing it once they look at what modes they have to choose from.