Add ~ to the arguments of genversion.bat
genversion.bat would fail if an argument was passed with quotations.
This is because the quotations were not being stripped before using
them as variables, causing the _temp files not being created.

Adding ~ to the arguments causes windows to strip the quotations,
enabling it to be used with path names containing spaces and the
MSVS pre-build event.

Signed-off-by: Ralph Giles <giles@mozilla.com>