VS2015: join common options in a properties file
Should make the projects a bit easier to digest and the differences
between the different configurations and platforms be more obvious.

A (happy?) side effect of this is the static libraries are now about
the same size as the ones produced by GCC and seem to build fine with
opus-tools.

Needs testing with other third party applications to make sure nothing
is missing that shouldn't.