Adds OPUS_{GS}ET_WAVEFORM_MATCHING encoder ctl
This option disables psychoacoustic improvements that deliberately alter the
waveform (e.g. changing the phase) to improve quality. Although it should
never be used for normal mono or stereo recordings, it can be useful for
cases where matching the waveform is important, e.g. direct ambisonics
(family 2), where the phase changes the spatial location of sounds.