cmake/CFeatureCheck.cmake: fix feature tests failing when Opus is a submodule
CMAKE_SOURCE_DIR corresponds to the top project's source directory.
CMAKE_BINARY_DIR corresponds to the top project's binary directory.

The usage of these variables doesn't cause any problems when Opus is built as a standalone project.

This is not the case when Opus is added as submodule: the variables are set by the project that calls "add_subdirectory()".

The fix consists in using PROJECT_SOURCE_DIR and PROJECT_BINARY_DIR, which always refer to the current project.