CFeatureCheck: Use CMAKE_CURRENT_SOURCE_DIR, not CMAKE_SOURCE_DIR
The latter gets the directory of the outermost CMakeLists.txt. In the case where
libopus is submoduled into another project and libopus is add_subdirectory'd, this
means feature check will look in the parent project instead of this project's 'cmake'
folder for a feature check c file. Changing to the former fixes this.