gitlab-ci: Re-enable builds for merge requests.
When the whitespace job was added, we marked it with

    only: merge_requests

because we wanted it to only run on new merge requests.

However, adding an `only` clause like this disables all the
unmarked jobs for merge requests, which means we don't get
the more important feedback about build failures.

Instead, use the `rules` key, added in Gitlab 12.3. This
more flexible mechanism is currently recommended and lets
us specify what we want without having to add additional
decorations to every other job.