Update Opus in ISOBMFF spec to version 1.0.0.
The version in the most recent draft has shipped in Firefox,
Chromium, and FFmpeg for several years and is quite stable.

Normally we would consider bumping the version field for the
dOps box, however as far as I know no incompatible version of
the box has ever been deployed, so I propose leaving it as is.

Signed-off-by: Jean-Marc Valin <jmvalin@jmvalin.ca>