Do not run valgrind on -ffast-math builds
valgrind does not properly handle the Denormals-Are-Zero (DAZ) flag,
which can be set by -ffast-math. That would not generally be a problem
except when the compiler implements sqrt(x) as x*rsqrt(x) with an
exception for x=0. In that case, when x is a non-zero denorm value,
the exception does not trigger with valgrind (but does on real
hardware), which causes the sqrt to evaluate to -inf instead of 0.