github actions: Run compat test on ubuntu 22.04
The `CMakeVersionTest` job tests building with
a fresh install of cmake v3.16 from upstream on
github's ubuntu 20.04-based runner image. With
the 20.04 LTS release now in extended support,
github no longer provides this image and the
test was not running.

Instead, try running the backward-compatible
test on the ubuntu 22.04-based image.

Signed-off-by: Jean-Marc Valin <jmvalin@jmvalin.ca>